rc.lua: Commit a failsafe, move the real rc.lua to awesome.lua
